refactor: move reaction reads to comment/api/reaction
