+++ apps/web/src/shared/api/supabaseReads.ts
@@ -12,7 +12,6 @@ import { getSupabaseClient, isNetworkError, SupabaseNetworkError } from './supab
 import type { Board } from '@/board/model/Board';
 import type { Comment } from '@/comment/model/Comment';
 import type { Reply } from '@/comment/model/Reply';
-import type { Reaction } from '@/comment/model/Reaction';
 import type { User } from '@/user/model/User';
 import { createTimestamp } from '@/shared/model/Timestamp';
 
@@ -788,89 +787,6 @@ export async function fetchReplyByIdFromSupabase(
     userProfileImage: data.user_profile_image || '',
     createdAt: createTimestamp(new Date(data.created_at)),
   };
-}
-
-// --- Reactions ---
-
-/**
- * Fetch reactions for a comment or reply.
- * Replaces: getReactions in reaction.ts
- * Uses index: idx_reactions_comment or idx_reactions_reply
- */
-export async function fetchReactionsFromSupabase(params: {
-  commentId: string;
-  replyId?: string;
-}): Promise<Reaction[]> {
-  const supabase = getSupabaseClient();
-
-  let q = supabase
-    .from('reactions')
-    .select('id, reaction_type, user_id, user_name, user_profile_image, created_at');
-
-  if (params.replyId) {
-    q = q.eq('reply_id', params.replyId);
-  } else {
-    q = q.eq('comment_id', params.commentId);
-  }
-
-  const { data, error } = await q;
-
-  if (error) {
-    console.error('Supabase fetchReactions error:', error);
-    throw error;
-  }
-
-  return (data || []).map((row) => ({
-    id: row.id,
-    content: row.reaction_type,
-    createdAt: new Date(row.created_at),
-    reactionUser: {
-      userId: row.user_id,
-      userName: row.user_name,
-      userProfileImage: row.user_profile_image || '',
-    },
-  }));
-}
-
-/**
- * Batch-fetch reactions for multiple comments in one query.
- * Returns Map<commentId, Reaction[]> for cache seeding.
- */
-export async function fetchBatchReactionsForComments(
-  commentIds: string[],
-): Promise<Map<string, Reaction[]>> {
-  if (commentIds.length === 0) return new Map();
-  const supabase = getSupabaseClient();
-
-  const { data, error } = await supabase
-    .from('reactions')
-    .select('id, comment_id, reaction_type, user_id, user_name, user_profile_image, created_at')
-    .in('comment_id', commentIds);
-
-  if (error) {
-    console.error('Supabase batch reactions fetch error:', { commentCount: commentIds.length, error });
-    throw error;
-  }
-
-  const result = new Map<string, Reaction[]>();
-  for (const commentId of commentIds) {
-    result.set(commentId, []);
-  }
-  for (const row of data || []) {
-    const reactions = result.get(row.comment_id) ?? [];
-    reactions.push({
-      id: row.id,
-      content: row.reaction_type,
-      createdAt: new Date(row.created_at),
-      reactionUser: {
-        userId: row.user_id,
-        userName: row.user_name,
-        userProfileImage: row.user_profile_image || '',
-      },
-    });
-    result.set(row.comment_id, reactions);
-  }
-  return result;
 }
 
 // --- Batch Queries (shared by stats + board pages) ---
